de-dupe, remove internal
